requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
jinja2==3.1.4
feedparser==6.0.11
requests==2.32.3
python-multipart==0.0.9
python-dotenv==1.0.0
